+++ tests/test_html_renderer_coverage.py
@@ -1,0 +1,181 @@
+"""HTMLRenderer Coverage Tests - メインレンダラーテスト
+
+HTMLRenderer完全テスト
+Target: main_renderer.py (45.05% → 完全)
+Goal: HTMLRenderer機能の完全カバレッジ
+"""
+
+import tempfile
+from pathlib import Path
+from unittest.mock import MagicMock, Mock, patch
+
+import pytest
+
+from kumihan_formatter.core.ast_nodes import Node
+from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
+
+
+class TestHTMLRenderer:
+    """HTMLRenderer完全テスト"""
+
+    def setup_method(self):
+        """テストセットアップ"""
+        self.renderer = HTMLRenderer()
+
+    def test_html_renderer_initialization(self):
+        """HTMLRenderer初期化テスト"""
+        renderer = HTMLRenderer()
+
+        # 必要な属性が初期化されていることを確認
+        assert hasattr(renderer, "element_renderer")
+        assert hasattr(renderer, "heading_renderer")
+        assert hasattr(renderer, "formatter")
+        assert hasattr(renderer, "content_processor")
+
+    def test_html_renderer_components_integration(self):
+        """HTMLRendererコンポーネント統合テスト"""
+        renderer = HTMLRenderer()
+
+        # 各コンポーネントが正しく初期化されていることを確認
+        assert renderer.element_renderer is not None
+        assert renderer.heading_renderer is not None
+        assert renderer.formatter is not None
+        assert renderer.content_processor is not None
+
+    def test_render_empty_ast(self):
+        """空のAST描画テスト"""
+        empty_ast = []
+        result = self.renderer.render_nodes(empty_ast)
+
+        # 空のASTでも正常に処理されることを確認
+        assert isinstance(result, str)
+        assert len(result) >= 0
+
+    def test_render_single_node(self):
+        """単一ノード描画テスト"""
+        # モックノードを作成
+        mock_node = Mock(spec=Node)
+        mock_node.type = "paragraph"
+        mock_node.content = "Test content"
+        mock_node.children = []
+        mock_node.attributes = {}
+
+        ast = [mock_node]
+
+        # render_nodeメソッドをテスト
+        result = self.renderer.render_node(mock_node)
+
+        # 正常にレンダリングされることを確認
+        assert isinstance(result, str)
+
+    def test_render_multiple_nodes(self):
+        """複数ノード描画テスト"""
+        # 複数のモックノードを作成
+        mock_nodes = []
+        for i in range(3):
+            node = Mock(spec=Node)
+            node.type = "paragraph"
+            node.content = f"Content {i}"
+            node.children = []
+            node.attributes = {}
+            mock_nodes.append(node)
+
+        result = self.renderer.render_nodes(mock_nodes)
+
+        # 複数ノードが処理されることを確認
+        assert isinstance(result, str)
+
+    def test_render_nested_nodes(self):
+        """ネストしたノード描画テスト"""
+        # ネストしたノード構造を作成
+        child_node = Mock(spec=Node)
+        child_node.type = "text"
+        child_node.content = "Child content"
+        child_node.children = []
+        child_node.attributes = {}
+
+        parent_node = Mock(spec=Node)
+        parent_node.type = "paragraph"
+        parent_node.content = "Parent content"
+        parent_node.children = [child_node]
+        parent_node.attributes = {}
+
+        ast = [parent_node]
+
+        result = self.renderer.render_nodes(ast)
+
+        # ネスト構造が処理されることを確認
+        assert isinstance(result, str)
+
+    def test_render_with_heading_collection(self):
+        """見出し収集付き描画テスト"""
+        # 見出しノードを作成
+        heading_node = Mock(spec=Node)
+        heading_node.type = "h1"  # heading ではなく h1 を使用
+        heading_node.content = "Test Heading"
+        heading_node.children = []
+        heading_node.attributes = {"level": "1"}
+
+        ast = [heading_node]
+
+        result = self.renderer.render_nodes(ast)
+
+        # 見出し収集が実行されることを確認
+        assert isinstance(result, str)
+
+    def test_render_with_formatting(self):
+        """フォーマット付き描画テスト"""
+        mock_node = Mock(spec=Node)
+        mock_node.type = "paragraph"
+        mock_node.content = "Test content"
+        mock_node.children = []
+        mock_node.attributes = {}
+
+        ast = [mock_node]
+
+        result = self.renderer.render_nodes(ast)
+
+        # フォーマッターが呼ばれることを確認
+        assert isinstance(result, str)
+
+    def test_render_error_handling(self):
+        """描画エラーハンドリングテスト"""
+        # 無効なノードでエラーを発生させる
+        invalid_node = Mock(spec=Node)
+        invalid_node.type = "invalid"
+        invalid_node.content = None
+        invalid_node.children = None
+        invalid_node.attributes = None
+
+        ast = [invalid_node]
+
+        # エラーが適切に処理されることを確認
+        try:
+            result = self.renderer.render_nodes(ast)
+            # エラーが発生しても何らかの結果が返されることを確認
+            assert isinstance(result, str)
+        except Exception:
+            # または例外が適切に処理されることを確認
+            assert True
+
+    def test_render_performance_with_large_ast(self):
+        """大きなAST描画パフォーマンステスト"""
+        # 大量のノードを作成
+        large_ast = []
+        for i in range(100):
+            node = Mock(spec=Node)
+            node.type = "paragraph"
+            node.content = f"Content {i}"
+            node.children = []
+            node.attributes = {}
+            large_ast.append(node)
+
+        import time
+
+        start = time.time()
+        result = self.renderer.render_nodes(large_ast)
+        end = time.time()
+
+        # 合理的な時間内で処理が完了することを確認
+        assert isinstance(result, str)
+        assert (end - start) < 5.0  # 5秒以内
