+++ tests/test_logger_integration_coverage.py
@@ -1,0 +1,79 @@
+"""Logger Integration Coverage Tests - ログシステム統合テスト
+
+LoggerIntegration完全テスト
+Target: config_manager_env.py + logger.py統合機能
+Goal: 設定システムとログシステムの統合カバレッジ
+"""
+
+import os
+import time
+from unittest.mock import patch
+
+from kumihan_formatter.config.base_config import BaseConfig
+from kumihan_formatter.config.config_manager_env import ConfigEnvironmentHandler
+from kumihan_formatter.core.utilities.logger import get_logger
+
+
+class TestLoggerIntegration:
+    """ロガー統合テスト"""
+
+    def test_config_and_logger_integration(self):
+        """設定とロガーの統合テスト"""
+        # 環境変数で設定を指定
+        with patch.dict(
+            os.environ,
+            {
+                "KUMIHAN_DEBUG": "true",
+                "KUMIHAN_LOG_LEVEL": "DEBUG",
+                "KUMIHAN_DEV_LOG": "true",
+            },
+        ):
+            # 設定読み込み
+            handler = ConfigEnvironmentHandler()
+            config = BaseConfig()
+            handler.load_from_env(config)
+
+            # ロガー取得
+            logger = get_logger("integration_test")
+            logger.info("Integration test message")
+
+            assert True
+
+    def test_logger_performance_with_config(self):
+        """設定連携パフォーマンステスト"""
+        with patch.dict(
+            os.environ, {"KUMIHAN_PERFORMANCE_LOG": "true", "KUMIHAN_LOG_LEVEL": "INFO"}
+        ):
+            logger = get_logger("performance_test")
+
+            start = time.time()
+
+            for i in range(10):
+                logger.info(f"Performance test message {i}")
+
+            end = time.time()
+            duration = end - start
+
+            # パフォーマンスログが正常に動作することを確認
+            logger.info("Performance test completed in %.3f seconds", duration)
+            assert duration < 1.0  # 1秒以内で完了
+
+    def test_error_handling_integration(self):
+        """エラーハンドリング統合テスト"""
+        logger = get_logger("error_integration_test")
+
+        # 様々なエラーシナリオでのログテスト
+        test_scenarios = [
+            lambda: logger.info(None),  # None value
+            lambda: logger.info(""),  # Empty string
+            lambda: logger.info("Test with unicode: 日本語"),  # Unicode
+            lambda: logger.info("Test with special chars: !@#$%^&*()"),  # Special chars
+        ]
+
+        for scenario in test_scenarios:
+            try:
+                scenario()
+            except Exception as e:
+                logger.error(f"Scenario failed: {e}")
+
+        assert True
