--- tests/test_phase1_18percent_achievement.py
@@ -1,381 +1,0 @@
-"""Phase 1 18%完全達成テスト - 最終アタック
-
-Current: 16.94% → Target: 18.00% (残り+1.06pt)
-Strategy: 中規模ファイルの効率的攻略
-Focus: config系, error_handling系, ファイル操作系
-"""
-
-from unittest.mock import Mock, patch
-
-import pytest
-
-
-class TestConfigSystemCoverage:
-    """config系システム カバレッジ向上"""
-
-    def test_base_config_import_coverage(self):
-        """base_config.py import カバーテスト"""
-        from kumihan_formatter.config.base_config import BaseConfig
-
-        assert BaseConfig is not None
-
-    def test_base_config_initialization(self):
-        """BaseConfig初期化テスト"""
-        from kumihan_formatter.config.base_config import BaseConfig
-
-        try:
-            config = BaseConfig()
-            assert config is not None
-        except Exception:
-            # 初期化エラーも想定内
-            assert True
-
-    def test_config_manager_basic_ops(self):
-        """config_manager.py 基本操作テスト"""
-        try:
-            from kumihan_formatter.config.config_manager import ConfigManager
-
-            # 基本クラス確認
-            assert ConfigManager is not None
-            assert hasattr(ConfigManager, "__init__")
-        except ImportError:
-            pytest.skip("ConfigManager import failed")
-
-    def test_config_manager_factory_functions(self):
-        """config_manager ファクトリ関数テスト"""
-        try:
-            from kumihan_formatter.config.config_manager import create_config_manager
-
-            result = create_config_manager()
-            assert result is not None or result is None  # どちらでも有効
-        except ImportError:
-            pytest.skip("create_config_manager import failed")
-        except Exception:
-            # 実行エラーも想定内
-            assert True
-
-    def test_config_manager_utils_coverage(self):
-        """config_manager_utils.py カバーテスト"""
-        try:
-            import kumihan_formatter.config.config_manager_utils as utils_module
-
-            # モジュール基本確認
-            assert utils_module is not None
-
-            # 利用可能な関数発見
-            public_items = [
-                item for item in dir(utils_module) if not item.startswith("_")
-            ]
-            assert len(public_items) >= 0
-        except ImportError:
-            pytest.skip("config_manager_utils import failed")
-
-    def test_extended_config_coverage(self):
-        """extended_config.py カバーテスト"""
-        try:
-            from kumihan_formatter.config.extended_config import ExtendedConfig
-
-            # 基本クラス確認
-            assert ExtendedConfig is not None
-        except ImportError:
-            pytest.skip("ExtendedConfig import failed")
-
-
-class TestErrorHandlingCoverage:
-    """error_handling系 カバレッジ向上"""
-
-    def test_error_factories_coverage(self):
-        """error_factories.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.error_handling import error_factories
-
-            assert error_factories is not None
-
-            # モジュール内の関数発見
-            functions = [
-                item
-                for item in dir(error_factories)
-                if not item.startswith("_") and callable(getattr(error_factories, item))
-            ]
-            assert len(functions) >= 0
-        except ImportError:
-            pytest.skip("error_factories import failed")
-
-    def test_error_recovery_coverage(self):
-        """error_recovery.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.error_handling import error_recovery
-
-            assert error_recovery is not None
-
-            # 基本的な関数テスト
-            public_items = [
-                item for item in dir(error_recovery) if not item.startswith("_")
-            ]
-            assert len(public_items) >= 0
-        except ImportError:
-            pytest.skip("error_recovery import failed")
-
-    def test_unified_handler_coverage(self):
-        """unified_handler.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.error_handling.unified_handler import (
-                UnifiedHandler,
-            )
-
-            assert UnifiedHandler is not None
-
-            # 基本初期化テスト
-            try:
-                handler = UnifiedHandler()
-                assert handler is not None
-            except Exception:
-                # 初期化エラーも想定内
-                assert True
-        except ImportError:
-            pytest.skip("UnifiedHandler import failed")
-
-    def test_context_manager_coverage(self):
-        """context_manager.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.error_handling.context_manager import (
-                ContextManager,
-            )
-
-            assert ContextManager is not None
-        except ImportError:
-            pytest.skip("ContextManager import failed")
-
-
-class TestFileOperationsCoverage:
-    """file_operations系 カバレッジ向上"""
-
-    def test_file_io_handler_coverage(self):
-        """file_io_handler.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.file_io_handler import FileIOHandler
-
-            assert FileIOHandler is not None
-
-            # 基本初期化テスト
-            try:
-                handler = FileIOHandler()
-                assert handler is not None
-            except Exception:
-                assert True
-        except ImportError:
-            pytest.skip("FileIOHandler import failed")
-
-    def test_file_operations_core_coverage(self):
-        """file_operations_core.py カバーテスト"""
-        try:
-            import kumihan_formatter.core.file_operations_core as core_module
-
-            assert core_module is not None
-
-            # 利用可能な関数・クラス発見
-            public_items = [
-                item for item in dir(core_module) if not item.startswith("_")
-            ]
-            assert len(public_items) >= 0
-        except ImportError:
-            pytest.skip("file_operations_core import failed")
-
-    def test_file_operations_factory_coverage(self):
-        """file_operations_factory.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.file_operations_factory import (
-                create_file_operations,
-            )
-
-            result = create_file_operations()
-            assert result is not None or result is None
-        except ImportError:
-            pytest.skip("create_file_operations import failed")
-        except Exception:
-            assert True
-
-    def test_file_validators_coverage(self):
-        """file_validators.py カバーテスト"""
-        try:
-            import kumihan_formatter.core.file_validators as validators_module
-
-            assert validators_module is not None
-
-            # バリデーター関数発見
-            validators = [
-                item
-                for item in dir(validators_module)
-                if not item.startswith("_")
-                and callable(getattr(validators_module, item))
-            ]
-            assert len(validators) >= 0
-        except ImportError:
-            pytest.skip("file_validators import failed")
-
-
-class TestRenderingSystemCoverage:
-    """rendering系 カバレッジ向上"""
-
-    def test_element_renderer_coverage(self):
-        """element_renderer.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.rendering.element_renderer import (
-                ElementRenderer,
-            )
-
-            assert ElementRenderer is not None
-
-            try:
-                renderer = ElementRenderer()
-                assert renderer is not None
-            except Exception:
-                assert True
-        except ImportError:
-            pytest.skip("ElementRenderer import failed")
-
-    def test_main_renderer_coverage(self):
-        """main_renderer.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.rendering.main_renderer import MainRenderer
-
-            assert MainRenderer is not None
-        except ImportError:
-            pytest.skip("MainRenderer import failed")
-
-    def test_html_formatter_coverage(self):
-        """html_formatter.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.rendering.html_formatter import HTMLFormatter
-
-            assert HTMLFormatter is not None
-
-            try:
-                formatter = HTMLFormatter()
-                assert formatter is not None
-            except Exception:
-                assert True
-        except ImportError:
-            pytest.skip("HTMLFormatter import failed")
-
-    def test_template_manager_coverage(self):
-        """template_manager.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.template_manager import TemplateManager
-
-            assert TemplateManager is not None
-
-            try:
-                manager = TemplateManager()
-                assert manager is not None
-            except Exception:
-                assert True
-        except ImportError:
-            pytest.skip("TemplateManager import failed")
-
-
-class TestUtilitiesCoverage:
-    """utilities系 カバレッジ向上"""
-
-    def test_logger_integration_coverage(self):
-        """logger統合カバーテスト"""
-        try:
-            from kumihan_formatter.core.utilities.logger import get_logger
-
-            logger = get_logger(__name__)
-            assert logger is not None
-
-            # 基本ログメソッド呼び出し
-            logger.info("test info message")
-            logger.debug("test debug message")
-            logger.warning("test warning message")
-
-            assert True
-        except ImportError:
-            pytest.skip("logger import failed")
-
-    def test_converters_coverage(self):
-        """converters.py カバーテスト"""
-        try:
-            import kumihan_formatter.core.utilities.converters as converters_module
-
-            assert converters_module is not None
-
-            # 変換関数発見
-            converters = [
-                item
-                for item in dir(converters_module)
-                if not item.startswith("_")
-                and callable(getattr(converters_module, item))
-            ]
-            assert len(converters) >= 0
-        except ImportError:
-            pytest.skip("converters import failed")
-
-    def test_data_structures_coverage(self):
-        """data_structures.py カバーテスト"""
-        try:
-            import kumihan_formatter.core.utilities.data_structures as ds_module
-
-            assert ds_module is not None
-
-            # データ構造クラス発見
-            classes = [item for item in dir(ds_module) if not item.startswith("_")]
-            assert len(classes) >= 0
-        except ImportError:
-            pytest.skip("data_structures import failed")
-
-
-class TestASTNodesCoverage:
-    """ast_nodes系 カバレッジ向上"""
-
-    def test_node_builder_coverage(self):
-        """node_builder.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.ast_nodes.node_builder import NodeBuilder
-
-            assert NodeBuilder is not None
-
-            try:
-                builder = NodeBuilder()
-                assert builder is not None
-            except Exception:
-                assert True
-        except ImportError:
-            pytest.skip("NodeBuilder import failed")
-
-    def test_factories_coverage(self):
-        """factories.py カバーテスト"""
-        try:
-            import kumihan_formatter.core.ast_nodes.factories as factories_module
-
-            assert factories_module is not None
-
-            # ファクトリ関数発見
-            factories = [
-                item
-                for item in dir(factories_module)
-                if not item.startswith("_")
-                and callable(getattr(factories_module, item))
-            ]
-            assert len(factories) >= 0
-        except ImportError:
-            pytest.skip("factories import failed")
-
-    def test_node_coverage(self):
-        """node.py カバーテスト"""
-        try:
-            from kumihan_formatter.core.ast_nodes.node import Node
-
-            assert Node is not None
-
-            try:
-                # 基本ノード作成テスト
-                node = Node("p", content="test content")
-                assert node is not None
-                assert node.type == "p"
-                assert node.content == "test content"
-            except Exception:
-                assert True
-        except ImportError:
-            pytest.skip("Node import failed")
