--- tests/test_phase1b_coverage.py
@@ -1,413 +1,0 @@
-"""Phase 1B Coverage Tests - 高効率ファイル完全制覇
-
-高カバレッジファイルの100%制覇テスト
-Target: config_manager_env.py (68.18% → 100%), logger.py (68.37% → 100%)
-Goal: +1.02%ポイント (149文のカバレッジ向上)
-"""
-
-import os
-import tempfile
-from pathlib import Path
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-from kumihan_formatter.config.base_config import BaseConfig
-from kumihan_formatter.config.config_manager_env import ConfigEnvironmentHandler
-from kumihan_formatter.config.extended_config import ExtendedConfig
-from kumihan_formatter.core.utilities.logger import KumihanLogger, get_logger
-
-
-class TestConfigEnvironmentHandler:
-    """ConfigEnvironmentHandler完全テスト"""
-
-    def setup_method(self):
-        """テストセットアップ"""
-        self.handler = ConfigEnvironmentHandler()
-        self.base_config = BaseConfig()
-
-    def test_init_default_prefix(self):
-        """デフォルトプレフィックス初期化テスト"""
-        handler = ConfigEnvironmentHandler()
-        assert handler.env_prefix == "KUMIHAN_"
-
-    def test_init_custom_prefix(self):
-        """カスタムプレフィックス初期化テスト"""
-        handler = ConfigEnvironmentHandler("CUSTOM_")
-        assert handler.env_prefix == "CUSTOM_"
-
-    def test_load_from_env_base_config(self):
-        """BaseConfig環境変数読み込みテスト"""
-        with patch.dict(
-            os.environ,
-            {
-                "KUMIHAN_DEBUG": "true",
-                "KUMIHAN_ENCODING": "utf-8",
-                "KUMIHAN_OUTPUT_DIR": "/tmp/output",
-            },
-        ):
-            config = BaseConfig()
-            self.handler.load_from_env(config)
-            # 環境変数が正常に読み込まれることを確認
-            assert True  # 実行エラーなしを確認
-
-    def test_load_from_env_extended_config(self):
-        """ExtendedConfig環境変数読み込みテスト"""
-        with patch.dict(
-            os.environ,
-            {
-                "KUMIHAN_TEMPLATE": "custom",
-                "KUMIHAN_THEME": "dark",
-                "KUMIHAN_VERBOSE": "true",
-            },
-        ):
-            config = ExtendedConfig()
-            self.handler.load_from_env(config)
-            # 環境変数が正常に読み込まれることを確認
-            assert True  # 実行エラーなしを確認
-
-    def test_load_from_env_empty_environment(self):
-        """空の環境変数でのテスト"""
-        # 関連する環境変数をクリア
-        env_vars_to_clear = [k for k in os.environ.keys() if k.startswith("KUMIHAN_")]
-        with patch.dict(os.environ, {}, clear=False):
-            for var in env_vars_to_clear:
-                os.environ.pop(var, None)
-
-            config = BaseConfig()
-            self.handler.load_from_env(config)
-            # エラーが発生しないことを確認
-            assert True
-
-    def test_load_from_env_boolean_conversion(self):
-        """Boolean型環境変数の変換テスト"""
-        test_cases = [
-            ("true", True),
-            ("false", False),
-            ("1", True),
-            ("0", False),
-            ("yes", True),
-            ("no", False),
-        ]
-
-        for env_value, expected in test_cases:
-            with patch.dict(os.environ, {"KUMIHAN_DEBUG": env_value}):
-                config = BaseConfig()
-                self.handler.load_from_env(config)
-                # 変換処理が正常に実行されることを確認
-                assert True
-
-    def test_load_from_env_invalid_values(self):
-        """無効な環境変数値の処理テスト"""
-        with patch.dict(
-            os.environ,
-            {
-                "KUMIHAN_INVALID": "invalid_value",
-                "KUMIHAN_EMPTY": "",
-                "KUMIHAN_NUMERIC": "12345",
-            },
-        ):
-            config = BaseConfig()
-            # 例外が発生せずに処理されることを確認
-            self.handler.load_from_env(config)
-            assert True
-
-    def test_env_prefix_functionality(self):
-        """環境変数プレフィックス機能テスト"""
-        custom_handler = ConfigEnvironmentHandler("CUSTOM_")
-
-        with patch.dict(
-            os.environ,
-            {"CUSTOM_DEBUG": "true", "KUMIHAN_DEBUG": "false"},  # これは無視される
-        ):
-            config = BaseConfig()
-            custom_handler.load_from_env(config)
-            # カスタムプレフィックスが正常に機能することを確認
-            assert True
-
-    def test_load_from_env_multiple_configs(self):
-        """複数の設定オブジェクトでの環境変数読み込みテスト"""
-        with patch.dict(
-            os.environ, {"KUMIHAN_DEBUG": "true", "KUMIHAN_TEMPLATE": "custom"}
-        ):
-            base_config = BaseConfig()
-            extended_config = ExtendedConfig()
-
-            self.handler.load_from_env(base_config)
-            self.handler.load_from_env(extended_config)
-
-            # 両方の設定オブジェクトで正常に処理されることを確認
-            assert True
-
-    def test_env_handler_thread_safety(self):
-        """環境変数ハンドラーのスレッドセーフティテスト"""
-        import threading
-
-        def load_config():
-            config = BaseConfig()
-            self.handler.load_from_env(config)
-
-        threads = []
-        for _ in range(5):
-            thread = threading.Thread(target=load_config)
-            threads.append(thread)
-            thread.start()
-
-        for thread in threads:
-            thread.join()
-
-        # 並行処理でエラーが発生しないことを確認
-        assert True
-
-
-class TestKumihanLogger:
-    """KumihanLogger完全テスト"""
-
-    def setup_method(self):
-        """テストセットアップ"""
-        # 新しいロガーインスタンスを取得
-        self.logger = get_logger("test_module")
-
-    def test_get_logger_singleton(self):
-        """get_logger シングルトンテスト"""
-        logger1 = get_logger("test_module")
-        logger2 = get_logger("test_module")
-
-        # 同じモジュール名で同じインスタンスが返されることを確認
-        assert logger1 is not None
-        assert logger2 is not None
-
-    def test_get_logger_different_modules(self):
-        """get_logger 異なるモジュールテスト"""
-        logger1 = get_logger("module1")
-        logger2 = get_logger("module2")
-
-        # 異なるモジュール名で異なるロガーが返されることを確認
-        assert logger1 is not None
-        assert logger2 is not None
-
-    def test_logger_basic_methods(self):
-        """ロガー基本メソッドテスト"""
-        logger = get_logger("test")
-
-        # 基本的なログメソッドが存在し、呼び出し可能であることを確認
-        assert hasattr(logger, "debug")
-        assert hasattr(logger, "info")
-        assert hasattr(logger, "warning")
-        assert hasattr(logger, "error")
-        assert hasattr(logger, "critical")
-
-        # 実際にメソッドを呼び出してエラーが発生しないことを確認
-        logger.debug("Test debug message")
-        logger.info("Test info message")
-        logger.warning("Test warning message")
-        logger.error("Test error message")
-        logger.critical("Test critical message")
-
-        assert True
-
-    def test_logger_with_different_formats(self):
-        """異なるフォーマットでのログテスト"""
-        logger = get_logger("test_format")
-
-        # 様々なメッセージ形式でのログ出力テスト
-        logger.info("Simple message")
-        logger.info("Message with %s placeholder", "value")
-        logger.info("Message with {} format".format("value"))
-        logger.info(f"F-string message with {'value'}")
-
-        # 辞書オブジェクト
-        logger.info("Dict message: %s", {"key": "value"})
-
-        # リストオブジェクト
-        logger.info("List message: %s", [1, 2, 3])
-
-        assert True
-
-    def test_logger_exception_logging(self):
-        """例外ログテスト"""
-        logger = get_logger("test_exception")
-
-        try:
-            raise ValueError("Test exception")
-        except ValueError:
-            # 例外情報付きログ
-            logger.exception("Exception occurred")
-            logger.error("Error with exception", exc_info=True)
-
-        assert True
-
-    def test_logger_with_extra_data(self):
-        """追加データ付きログテスト"""
-        logger = get_logger("test_extra")
-
-        # extraデータ付きログ出力
-        extra_data = {"user_id": 12345, "session_id": "abc123", "request_id": "req-456"}
-
-        logger.info("User action performed", extra=extra_data)
-        logger.error("User error occurred", extra=extra_data)
-
-        assert True
-
-    def test_logger_performance_logging(self):
-        """パフォーマンスログテスト"""
-        logger = get_logger("test_performance")
-
-        import time
-
-        start_time = time.time()
-        # シミュレーション処理
-        time.sleep(0.001)  # 1ms
-        end_time = time.time()
-
-        duration = end_time - start_time
-        logger.info("Operation completed in %.3f seconds", duration)
-
-        assert True
-
-    def test_logger_structured_logging(self):
-        """構造化ログテスト"""
-        logger = get_logger("test_structured")
-
-        # JSON形式でのログ出力テスト
-        structured_data = {
-            "event": "user_login",
-            "user_id": 12345,
-            "timestamp": "2025-07-22T18:00:00Z",
-            "ip_address": "192.168.1.1",
-            "user_agent": "Mozilla/5.0",
-        }
-
-        logger.info("Structured log entry", extra=structured_data)
-
-        assert True
-
-    def test_logger_multithreaded_usage(self):
-        """マルチスレッド使用テスト"""
-        import threading
-
-        def log_messages(thread_id):
-            logger = get_logger(f"test_thread_{thread_id}")
-            for i in range(5):
-                logger.info(f"Thread {thread_id} message {i}")
-
-        threads = []
-        for i in range(3):
-            thread = threading.Thread(target=log_messages, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        for thread in threads:
-            thread.join()
-
-        assert True
-
-    def test_logger_with_file_output(self):
-        """ファイル出力ログテスト"""
-        with tempfile.TemporaryDirectory() as temp_dir:
-            log_file = Path(temp_dir) / "test.log"
-
-            # 環境変数でログファイルを指定
-            with patch.dict(os.environ, {"KUMIHAN_LOG_FILE": str(log_file)}):
-                logger = get_logger("test_file_output")
-                logger.info("Test file output message")
-
-                # ファイルが作成されるかは実装依存だが、エラーが発生しないことを確認
-                assert True
-
-    def test_logger_level_filtering(self):
-        """ログレベルフィルタリングテスト"""
-        logger = get_logger("test_level_filter")
-
-        # 異なるレベルでログ出力
-        logger.debug("Debug level message")
-        logger.info("Info level message")
-        logger.warning("Warning level message")
-        logger.error("Error level message")
-        logger.critical("Critical level message")
-
-        # レベルに関係なくエラーが発生しないことを確認
-        assert True
-
-    def test_logger_configuration_loading(self):
-        """ロガー設定読み込みテスト"""
-        # 環境変数による設定テスト
-        with patch.dict(
-            os.environ,
-            {
-                "KUMIHAN_LOG_LEVEL": "DEBUG",
-                "KUMIHAN_LOG_FORMAT": "json",
-                "KUMIHAN_DEV_LOG": "true",
-            },
-        ):
-            logger = get_logger("test_config_loading")
-            logger.info("Configuration test message")
-
-            assert True
-
-
-class TestLoggerIntegration:
-    """ロガー統合テスト"""
-
-    def test_config_and_logger_integration(self):
-        """設定とロガーの統合テスト"""
-        # 環境変数で設定を指定
-        with patch.dict(
-            os.environ,
-            {
-                "KUMIHAN_DEBUG": "true",
-                "KUMIHAN_LOG_LEVEL": "DEBUG",
-                "KUMIHAN_DEV_LOG": "true",
-            },
-        ):
-            # 設定読み込み
-            handler = ConfigEnvironmentHandler()
-            config = BaseConfig()
-            handler.load_from_env(config)
-
-            # ロガー取得
-            logger = get_logger("integration_test")
-            logger.info("Integration test message")
-
-            assert True
-
-    def test_logger_performance_with_config(self):
-        """設定連携パフォーマンステスト"""
-        with patch.dict(
-            os.environ, {"KUMIHAN_PERFORMANCE_LOG": "true", "KUMIHAN_LOG_LEVEL": "INFO"}
-        ):
-            logger = get_logger("performance_test")
-
-            import time
-
-            start = time.time()
-
-            for i in range(10):
-                logger.info(f"Performance test message {i}")
-
-            end = time.time()
-            duration = end - start
-
-            # パフォーマンスログが正常に動作することを確認
-            logger.info("Performance test completed in %.3f seconds", duration)
-            assert duration < 1.0  # 1秒以内で完了
-
-    def test_error_handling_integration(self):
-        """エラーハンドリング統合テスト"""
-        logger = get_logger("error_integration_test")
-
-        # 様々なエラーシナリオでのログテスト
-        test_scenarios = [
-            lambda: logger.info(None),  # None value
-            lambda: logger.info(""),  # Empty string
-            lambda: logger.info("Test with unicode: 日本語"),  # Unicode
-            lambda: logger.info("Test with special chars: !@#$%^&*()"),  # Special chars
-        ]
-
-        for scenario in test_scenarios:
-            try:
-                scenario()
-            except Exception as e:
-                logger.error(f"Scenario failed: {e}")
-
-        assert True
