--- tests/test_phase1c_coverage.py
@@ -1,533 +1,0 @@
-"""Phase 1C Coverage Tests - コア機能テスト拡充
-
-レンダリング・テンプレート機能の完全テスト
-Target: main_renderer.py (45.05% → 完全), element_renderer.py (38.96% → 完全), template_context.py (40.00% → 完全)
-Goal: +1.42%ポイント (205文のカバレッジ向上)
-"""
-
-import tempfile
-from pathlib import Path
-from unittest.mock import MagicMock, Mock, patch
-
-import pytest
-
-from kumihan_formatter.core.ast_nodes import Node
-from kumihan_formatter.core.rendering.element_renderer import ElementRenderer
-from kumihan_formatter.core.rendering.main_renderer import HTMLRenderer
-from kumihan_formatter.core.template_context import RenderContext
-
-
-class TestHTMLRenderer:
-    """HTMLRenderer完全テスト"""
-
-    def setup_method(self):
-        """テストセットアップ"""
-        self.renderer = HTMLRenderer()
-
-    def test_html_renderer_initialization(self):
-        """HTMLRenderer初期化テスト"""
-        renderer = HTMLRenderer()
-
-        # 必要な属性が初期化されていることを確認
-        assert hasattr(renderer, "element_renderer")
-        assert hasattr(renderer, "heading_renderer")
-        assert hasattr(renderer, "formatter")
-        assert hasattr(renderer, "content_processor")
-
-    def test_html_renderer_components_integration(self):
-        """HTMLRendererコンポーネント統合テスト"""
-        renderer = HTMLRenderer()
-
-        # 各コンポーネントが正しく初期化されていることを確認
-        assert renderer.element_renderer is not None
-        assert renderer.heading_renderer is not None
-        assert renderer.formatter is not None
-        assert renderer.content_processor is not None
-
-    def test_render_empty_ast(self):
-        """空のAST描画テスト"""
-        empty_ast = []
-        result = self.renderer.render_nodes(empty_ast)
-
-        # 空のASTでも正常に処理されることを確認
-        assert isinstance(result, str)
-        assert len(result) >= 0
-
-    def test_render_single_node(self):
-        """単一ノード描画テスト"""
-        # モックノードを作成
-        mock_node = Mock(spec=Node)
-        mock_node.type = "paragraph"
-        mock_node.content = "Test content"
-        mock_node.children = []
-        mock_node.attributes = {}
-
-        ast = [mock_node]
-
-        # render_nodeメソッドをテスト
-        result = self.renderer.render_node(mock_node)
-
-        # 正常にレンダリングされることを確認
-        assert isinstance(result, str)
-
-    def test_render_multiple_nodes(self):
-        """複数ノード描画テスト"""
-        # 複数のモックノードを作成
-        mock_nodes = []
-        for i in range(3):
-            node = Mock(spec=Node)
-            node.type = "paragraph"
-            node.content = f"Content {i}"
-            node.children = []
-            node.attributes = {}
-            mock_nodes.append(node)
-
-        result = self.renderer.render_nodes(mock_nodes)
-
-        # 複数ノードが処理されることを確認
-        assert isinstance(result, str)
-
-    def test_render_nested_nodes(self):
-        """ネストしたノード描画テスト"""
-        # ネストしたノード構造を作成
-        child_node = Mock(spec=Node)
-        child_node.type = "text"
-        child_node.content = "Child content"
-        child_node.children = []
-        child_node.attributes = {}
-
-        parent_node = Mock(spec=Node)
-        parent_node.type = "paragraph"
-        parent_node.content = "Parent content"
-        parent_node.children = [child_node]
-        parent_node.attributes = {}
-
-        ast = [parent_node]
-
-        result = self.renderer.render_nodes(ast)
-
-        # ネスト構造が処理されることを確認
-        assert isinstance(result, str)
-
-    def test_render_with_heading_collection(self):
-        """見出し収集付き描画テスト"""
-        # 見出しノードを作成
-        heading_node = Mock(spec=Node)
-        heading_node.type = "h1"  # heading ではなく h1 を使用
-        heading_node.content = "Test Heading"
-        heading_node.children = []
-        heading_node.attributes = {"level": "1"}
-
-        ast = [heading_node]
-
-        result = self.renderer.render_nodes(ast)
-
-        # 見出し収集が実行されることを確認
-        assert isinstance(result, str)
-
-    def test_render_with_formatting(self):
-        """フォーマット付き描画テスト"""
-        mock_node = Mock(spec=Node)
-        mock_node.type = "paragraph"
-        mock_node.content = "Test content"
-        mock_node.children = []
-        mock_node.attributes = {}
-
-        ast = [mock_node]
-
-        result = self.renderer.render_nodes(ast)
-
-        # フォーマッターが呼ばれることを確認
-        assert isinstance(result, str)
-
-    def test_render_error_handling(self):
-        """描画エラーハンドリングテスト"""
-        # 無効なノードでエラーを発生させる
-        invalid_node = Mock(spec=Node)
-        invalid_node.type = "invalid"
-        invalid_node.content = None
-        invalid_node.children = None
-        invalid_node.attributes = None
-
-        ast = [invalid_node]
-
-        # エラーが適切に処理されることを確認
-        try:
-            result = self.renderer.render_nodes(ast)
-            # エラーが発生しても何らかの結果が返されることを確認
-            assert isinstance(result, str)
-        except Exception:
-            # または例外が適切に処理されることを確認
-            assert True
-
-    def test_render_performance_with_large_ast(self):
-        """大きなAST描画パフォーマンステスト"""
-        # 大量のノードを作成
-        large_ast = []
-        for i in range(100):
-            node = Mock(spec=Node)
-            node.type = "paragraph"
-            node.content = f"Content {i}"
-            node.children = []
-            node.attributes = {}
-            large_ast.append(node)
-
-        import time
-
-        start = time.time()
-        result = self.renderer.render_nodes(large_ast)
-        end = time.time()
-
-        # 合理的な時間内で処理が完了することを確認
-        assert isinstance(result, str)
-        assert (end - start) < 5.0  # 5秒以内
-
-
-class TestElementRenderer:
-    """ElementRenderer完全テスト"""
-
-    def setup_method(self):
-        """テストセットアップ"""
-        self.renderer = ElementRenderer()
-
-    def test_element_renderer_initialization(self):
-        """ElementRenderer初期化テスト"""
-        renderer = ElementRenderer()
-
-        # 必要なコンポーネントレンダラーが初期化されていることを確認
-        assert hasattr(renderer, "basic_renderer")
-        assert hasattr(renderer, "heading_renderer")
-        assert hasattr(renderer, "list_renderer")
-        assert hasattr(renderer, "div_renderer")
-
-    def test_render_paragraph_element(self):
-        """段落要素レンダリングテスト"""
-        node = Mock(spec=Node)
-        node.type = "paragraph"
-        node.content = "Test paragraph"
-        node.children = []
-        node.attributes = {}
-
-        # 段落レンダリングメソッドを呼び出し
-        result = self.renderer.render_paragraph(node)
-
-        # 段落が正常にレンダリングされることを確認
-        assert isinstance(result, str)
-
-    def test_render_heading_element(self):
-        """見出し要素レンダリングテスト"""
-        node = Mock(spec=Node)
-        node.type = "heading"
-        node.content = "Test Heading"
-        node.children = []
-        node.attributes = {"level": "1"}
-
-        # 見出しレンダリングメソッドを呼び出し
-        result = self.renderer.render_heading(node, 1)
-
-        # 見出しが正常にレンダリングされることを確認
-        assert isinstance(result, str)
-
-    def test_render_list_element(self):
-        """リスト要素レンダリングテスト"""
-        node = Mock(spec=Node)
-        node.type = "list"
-        node.content = ""
-        node.children = []
-        node.attributes = {"list_type": "ul"}
-
-        # リストレンダリングメソッドを呼び出し
-        result = self.renderer.render_unordered_list(node)
-
-        # リストが正常にレンダリングされることを確認
-        assert isinstance(result, str)
-
-    def test_render_div_element(self):
-        """div要素レンダリングテスト"""
-        node = Mock(spec=Node)
-        node.type = "div"
-        node.content = "Div content"
-        node.children = []
-        node.attributes = {"class": "test-div"}
-
-        # divレンダリングメソッドを呼び出し
-        result = self.renderer.render_div(node)
-
-        # divが正常にレンダリングされることを確認
-        assert isinstance(result, str)
-
-    def test_render_unknown_element(self):
-        """未知の要素レンダリングテスト"""
-        node = Mock(spec=Node)
-        node.type = "unknown"
-        node.content = "Unknown content"
-        node.children = []
-        node.attributes = {}
-
-        # 未知の要素でも何らかの処理が行われることを確認
-        result = self.renderer.render_generic(node)
-        assert isinstance(result, str)
-
-    def test_render_element_with_context(self):
-        """コンテキスト付き要素レンダリングテスト"""
-        node = Mock(spec=Node)
-        node.type = "paragraph"
-        node.content = "Test content"
-        node.children = []
-        node.attributes = {}
-
-        context = {"template_name": "custom", "theme": "dark"}
-
-        # コンテキスト付きで正常にレンダリングされることを確認
-        result = self.renderer.render_paragraph(node)
-        assert isinstance(result, str)
-
-    def test_render_element_with_children(self):
-        """子要素付き要素レンダリングテスト"""
-        child_node = Mock(spec=Node)
-        child_node.type = "text"
-        child_node.content = "Child text"
-        child_node.children = []
-        child_node.attributes = {}
-
-        parent_node = Mock(spec=Node)
-        parent_node.type = "paragraph"
-        parent_node.content = ""
-        parent_node.children = [child_node]
-        parent_node.attributes = {}
-
-        # 子要素付きで正常にレンダリングされることを確認
-        result = self.renderer.render_paragraph(parent_node)
-        assert isinstance(result, str)
-
-    def test_renderer_component_integration(self):
-        """レンダラーコンポーネント統合テスト"""
-        # 各タイプの要素を順番にテスト
-        element_types = ["paragraph", "heading", "list", "div"]
-
-        for element_type in element_types:
-            node = Mock(spec=Node)
-            node.type = element_type
-            node.content = f"Test {element_type}"
-            node.children = []
-            node.attributes = {}
-
-            # 各タイプに応じたメソッドを呼び出し
-            if element_type == "paragraph":
-                result = self.renderer.render_paragraph(node)
-            elif element_type == "heading":
-                result = self.renderer.render_heading(node, 1)
-            elif element_type == "list":
-                result = self.renderer.render_unordered_list(node)
-            elif element_type == "div":
-                result = self.renderer.render_div(node)
-            else:
-                result = self.renderer.render_generic(node)
-
-            # 全ての要素タイプで何らかの結果が返されることを確認
-            assert isinstance(result, str)
-
-    def test_renderer_error_resilience(self):
-        """レンダラーエラー耐性テスト"""
-        # 様々な無効なノードでテスト
-        invalid_cases = [
-            {"type": None, "content": "test"},
-            {"type": "paragraph", "content": None},
-            {"type": "paragraph", "children": None},
-            {"type": "paragraph", "attributes": None},
-        ]
-
-        for case in invalid_cases:
-            node = Mock(spec=Node)
-            for attr, value in case.items():
-                setattr(node, attr, value)
-
-            # エラーが発生しても処理が継続されることを確認
-            try:
-                result = self.renderer.render_generic(node)
-                assert isinstance(result, str)
-            except Exception:
-                # 例外が発生しても適切に処理されることを確認
-                assert True
-
-
-class TestRenderContext:
-    """TemplateContext完全テスト"""
-
-    def setup_method(self):
-        """テストセットアップ"""
-        self.context = RenderContext()
-
-    def test_template_context_initialization(self):
-        """TemplateContext初期化テスト"""
-        context = RenderContext()
-
-        # 基本的な属性が初期化されていることを確認
-        assert hasattr(context, "_context")
-        assert isinstance(context._context, dict)
-
-    def test_set_context_value(self):
-        """コンテキスト値設定テスト"""
-        self.context.custom("title", "Test Title")
-        self.context.custom("author", "Test Author")
-
-        # 値が正しく設定されることを確認
-        assert self.context.get("title") == "Test Title"
-        assert self.context.get("author") == "Test Author"
-
-    def test_get_context_value(self):
-        """コンテキスト値取得テスト"""
-        # 存在する値の取得
-        self.context.custom("test_key", "test_value")
-        result = self.context.get("test_key")
-        assert result == "test_value"
-
-        # 存在しない値の取得（デフォルト値）
-        result = self.context.get("nonexistent_key", "default_value")
-        assert result == "default_value"
-
-        # 存在しない値の取得（デフォルトなし）
-        result = self.context.get("nonexistent_key")
-        assert result is None
-
-    def test_update_context(self):
-        """コンテキスト一括更新テスト"""
-        update_data = {
-            "title": "Updated Title",
-            "description": "Test Description",
-            "version": "1.0.0",
-        }
-
-        self.context.merge(update_data)
-
-        # 全ての値が正しく更新されることを確認
-        for key, value in update_data.items():
-            assert self.context.get(key) == value
-
-    def test_context_merge(self):
-        """コンテキストマージテスト"""
-        # 初期データ
-        self.context.custom("title", "Original Title")
-        self.context.custom("author", "Original Author")
-
-        # マージデータ
-        merge_data = {
-            "title": "New Title",  # 上書き
-            "description": "New Description",  # 新規追加
-        }
-
-        self.context.merge(merge_data)
-
-        # 上書きと新規追加が正しく行われることを確認
-        assert self.context.get("title") == "New Title"
-        assert self.context.get("author") == "Original Author"
-        assert self.context.get("description") == "New Description"
-
-    def test_context_with_nested_data(self):
-        """ネストしたデータのコンテキストテスト"""
-        nested_data = {
-            "config": {"theme": "dark", "language": "ja"},
-            "metadata": {"created_at": "2025-07-22", "tags": ["test", "coverage"]},
-        }
-
-        self.context.merge(nested_data)
-
-        # ネストしたデータが正しく格納されることを確認
-        config = self.context.get("config")
-        assert config["theme"] == "dark"
-        assert config["language"] == "ja"
-
-        metadata = self.context.get("metadata")
-        assert metadata["created_at"] == "2025-07-22"
-        assert "test" in metadata["tags"]
-
-    def test_context_serialization(self):
-        """コンテキストシリアライゼーションテスト"""
-        self.context.custom("title", "Test Title")
-        self.context.custom("count", 42)
-        self.context.custom("active", True)
-
-        # build メソッドでシリアライゼーションテスト
-        result = self.context.build()
-        assert isinstance(result, dict)
-        assert result["title"] == "Test Title"
-        assert result["count"] == 42
-        assert result["active"] is True
-
-    def test_context_clear(self):
-        """コンテキストクリアテスト"""
-        self.context.custom("temp_data", "to be cleared")
-
-        # clearメソッドがある場合のテスト
-        if hasattr(self.context, "clear"):
-            self.context.clear()
-            assert self.context.get("temp_data") is None
-
-    def test_context_key_existence(self):
-        """コンテキストキー存在確認テスト"""
-        self.context.custom("existing_key", "value")
-
-        # hasメソッドがある場合のテスト
-        if hasattr(self.context, "has"):
-            assert self.context.has("existing_key") is True
-            assert self.context.has("nonexistent_key") is False
-
-    def test_context_iteration(self):
-        """コンテキスト反復テスト"""
-        test_data = {"key1": "value1", "key2": "value2", "key3": "value3"}
-
-        self.context.merge(test_data)
-
-        # keysメソッドがある場合のテスト
-        if hasattr(self.context, "keys"):
-            keys = self.context.keys()
-            assert len(keys) >= len(test_data)
-
-            # 全てのキーが含まれていることを確認
-            for test_key in test_data.keys():
-                assert test_key in keys
-
-        # itemsは実装されていないのでkeysでテスト済み
-        assert True
-
-    def test_context_performance_with_large_data(self):
-        """大量データでのコンテキストパフォーマンステスト"""
-        # 大量のデータを設定
-        large_data = {}
-        for i in range(1000):
-            large_data[f"key_{i}"] = f"value_{i}"
-
-        import time
-
-        start = time.time()
-        self.context.merge(large_data)
-        end = time.time()
-
-        # 合理的な時間内で処理が完了することを確認
-        assert (end - start) < 1.0  # 1秒以内
-
-        # データが正しく設定されていることを確認
-        assert self.context.get("key_500") == "value_500"
-        assert self.context.get("key_999") == "value_999"
-
-    def test_context_thread_safety(self):
-        """コンテキストスレッドセーフティテスト"""
-        import threading
-
-        def update_context(thread_id):
-            for i in range(10):
-                self.context.custom(f"thread_{thread_id}_key_{i}", f"value_{i}")
-
-        threads = []
-        for i in range(5):
-            thread = threading.Thread(target=update_context, args=(i,))
-            threads.append(thread)
-            thread.start()
-
-        for thread in threads:
-            thread.join()
-
-        # スレッド競合でエラーが発生していないことを確認
-        # いくつかのキーが正しく設定されていることを確認
-        assert self.context.get("thread_0_key_0") == "value_0"
-        assert self.context.get("thread_4_key_9") == "value_9"
